% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPIThemes-package.R
\docType{package}
\name{MPIThemes-package}
\alias{MPIThemes}
\alias{MPIThemes-package}
\title{MPIThemes: Unofficial MAX PLANCK INSTITUTE RMarkdown templates and ggplot themes}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package contains Rmakdown templates for creating the report,
  presentation, and other academic templates and ggplot themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ehyaei/MPIThemes}
  \item Report bugs at \url{https://github.com/Ehyaei/MPIThemes/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Ehyaei \email{ahmad.ehyaei@gmail.com}

}
\keyword{internal}
