% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{palette_mpi}
\alias{palette_mpi}
\title{Max-Planck institute palette}
\usage{
palette_mpi()
}
\value{
a vector containing the colors.
}
\description{
This palette was inspired by the CD manual (available \href{https://docplayer.org/2328711-Max-planck-institut-das-erscheinungsbild-der-max-planck-gesellschaft-4-ueberarbeitete-auflage.html}{here} )
of the Max Planck Society.
}
\examples{
library(MPIThemes)
scales::show_col(palette_mpi())
}
