% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_mpi}
\alias{scale_color_mpi}
\title{Fill scale constructor for MPI palette}
\usage{
scale_color_mpi(...)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale()}
}
\value{

}
\description{
Fill scale constructor for MPI palette
}
\examples{
ggplot(data = mtcars, aes(x = wt, y= mpg,color = as.character(cyl))) +
  geom_point()+scale_color_mpi()
}
