% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_scientific}
\alias{theme_scientific}
\title{Theme 'scientific'}
\usage{
theme_scientific(base_size = 12)
}
\arguments{
\item{base_size}{base font size, given in pts.}
}
\value{
scientific theme
}
\description{
Theme 'scientific'
}
\examples{
library(ggplot2)
library(dplyr)
iris \%>\%
group_by(Species) \%>\%
  summarise(Sepal.Width = mean(Sepal.Width)) \%>\%
  ggplot(aes(x = Species, y = Sepal.Width, fill = Species)) +
  geom_col() +
  labs(x="Species",
       y="Average Sepal Width",
       fill = "Species",
       title = "Iris Dataset")+
  theme_scientific()+
  scale_fill_mpi()
}
